+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##### Breaking
 
 - SourceKit based indexing has been removed, the IndexStore indexer is now the sole indexer. Therefore, the following scan options have been removed: `--use-index-store`, `--use-build-log`, `--save-build-log`.
+- The `scan-syntax` command has been removed.
 
 ##### Enhancements